backend/pyproject.toml
[project]
name = "kobo-backend"
version = "0.1.0"
description = "KOBO Autonomous Execution Workspace backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiohttp>=3.12.15",
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "minio>=7.2.18",
  "passlib[bcrypt]>=1.7.4",
  "prisma>=0.15.0",
  "pypdf>=6.0.0",
  "pydantic>=2.11.7",
  "pydantic-settings>=2.10.1",
  "python-docx>=1.2.0",
  "python-jose[cryptography]>=3.5.0",
  "python-multipart>=0.0.20",
  "redis>=6.4.0",
  "tenacity>=9.1.2",
  "uvicorn[standard]>=0.35.0",
  "websockets>=15.0.1",
]

[dependency-groups]
dev = [
  "mypy>=1.17.1",
  "pytest>=8.4.1",
  "pytest-asyncio>=1.1.0",
  "pytest-cov>=6.2.1",
  "ruff>=0.12.8",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
warn_unused_ignores = false
warn_return_any = false
disallow_untyped_defs = false
disable_error_code = ["arg-type", "assignment", "return-value", "call-overload"]
plugins = []

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
